feat: implement comprehensive monitoring and observability system (v0.2)
🔍 Monitoring & Observability Features:
- Integrated Micrometer + Prometheus metrics collection across all services
- Added comprehensive business metrics for rate limiting operations
- Implemented Grafana dashboards for real-time system monitoring
- Added structured logging and distributed tracing support

📊 Key Metrics Added:
- Rate limit check counters (total, allowed, denied) with tenant/resource tags
- Processing path metrics (local vs Redis fallback)
- Request latency histograms (P50, P95, P99)
- Policy cache hit/miss ratios and cache size gauges
- Kafka event publishing success/failure rates
- JVM, database, and Redis performance metrics

🏗️ Infrastructure Enhancements:
- Docker Compose setup for Prometheus + Grafana stack
- Automated metrics collection from all microservices (: 8081, :8082, : 8083)
- Pre-configured Grafana dashboards for business and technical metrics
- Health check endpoints for service monitoring

⚡ Performance Improvements:
- Optimized Kafka batch processing with intelligent duplicate detection
- Enhanced error handling and retry mechanisms
- Improved Redis Lua script performance monitoring
- Reduced database queries through smart batch operations

🛠️ Technical Implementation:
- RateLimiterMetricsService for centralized metrics collection
- Custom tags for multi-dimensional metric analysis
- Timer-based latency measurements for performance optimization
- Integration with existing CheckUseCaseService without breaking changes

📈 Business Value:
- Real-time visibility into rate limiting effectiveness
- Tenant-level usage analytics and quota monitoring
- Performance bottleneck identification and capacity planning
- SLA compliance tracking and alerting capabilities

🧪 Testing & Validation:
- Comprehensive test script for metrics generation
- Multi-scenario testing (allowed, denied, policy-not-found cases)
- Load testing support for performance validation
- Grafana dashboard verification with real data

This release transforms the Rate Limiter Platform into a production-ready,
observable system with comprehensive monitoring capabilities.

+++ limiter-data-plane/src/main/java/com/ratelimiter/dataplane/infrastructure/persistence/redis/RedisRateLimiterRepositoryTest.java
@@ -5,6 +5,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
@@ -21,7 +23,7 @@ class RedisRateLimiterRepositoryTest {
                         "test_tenant",
                         "/test/api",
                         100L,
-                        10.0,
+                        BigDecimal.valueOf(10.00),
                         1L,
                         "test-request-" + System.currentTimeMillis(),
                         System.currentTimeMillis()
